requirements.txt
asgiref==3.4.1
Babel==2.9.1
Django==3.2.7
django-phonenumber-field==5.2.0
phonenumbers==8.12.31
pytz==2021.1
secure-smtplib==0.1.1
sqlparse==0.4.2
